+++ src/alfasim_sdk/_internal/alfacase/schema.py
@@ -36,6 +36,11 @@ casing_section_description_schema = Map(
         Optional("material_above_filler"): Str(),
     }
 )
+combined_fluid_description_schema = Map(
+    {
+        Optional("pvt_model"): Str(),
+    }
+)
 composition_description_schema = Map(
     {
         "component": Str(),
@@ -755,6 +760,12 @@ casing_description_schema = Map(
         Optional("open_holes"): Seq(open_hole_description_schema),
     }
 )
+compositional_fluid_description_schema = Map(
+    {
+        Optional("composition"): Seq(composition_description_schema),
+        Optional("fraction_pairs"): Seq(bip_description_schema),
+    }
+)
 compressor_equipment_description_schema = Map(
     {
         "position": Map({"value": Float(), "unit": Str()}),
@@ -786,12 +797,6 @@ environment_description_schema = Map(
         Optional("reference_y_coordinate"): Map({"value": Float(), "unit": Str()}),
         Optional("md_properties_table"): Seq(environment_property_description_schema),
         Optional("tvd_properties_table"): Seq(environment_property_description_schema),
-    }
-)
-fluid_description_schema = Map(
-    {
-        Optional("composition"): Seq(composition_description_schema),
-        Optional("fraction_pairs"): Seq(bip_description_schema),
     }
 )
 formation_description_schema = Map(
@@ -897,6 +902,12 @@ pump_equipment_description_schema = Map(
         Optional("flow_direction"): Enum(['forward', 'backward']),
     }
 )
+pvt_model_combined_description_schema = Map(
+    {
+        Optional("reference_pvt_model"): Str(),
+        Optional("fluids"): MapPattern(Str(), combined_fluid_description_schema),
+    }
+)
 table_ipr_description_schema = Map(
     {
         Optional("well_index_phase"): Enum(['well_index_phase_gas', 'well_index_phase_oil', 'well_index_phase_water', 'well_index_phase_liquid']),
@@ -986,7 +997,7 @@ pvt_model_compositional_description_schema = Map(
         Optional("viscosity_model"): Enum(['corresponding_states_principle', 'lohrenz_bray_clark']),
         Optional("heavy_components"): Seq(heavy_component_description_schema),
         Optional("light_components"): Seq(light_component_description_schema),
-        Optional("fluids"): MapPattern(Str(), fluid_description_schema),
+        Optional("fluids"): MapPattern(Str(), compositional_fluid_description_schema),
     }
 )
 trends_output_description_schema = Map(
@@ -1040,7 +1051,8 @@ pvt_models_description_schema = Map(
         Optional("default_model"): Str(),
         Optional("tables"): MapPattern(Str(), Str()),
         Optional("correlations"): MapPattern(Str(), pvt_model_correlation_description_schema),
-        Optional("compositions"): MapPattern(Str(), pvt_model_compositional_description_schema),
+        Optional("compositional"): MapPattern(Str(), pvt_model_compositional_description_schema),
+        Optional("combined"): MapPattern(Str(), pvt_model_combined_description_schema),
     }
 )
 well_description_schema = Map(
@@ -1076,5 +1088,5 @@ case_description_schema = Map(
         Optional("walls"): Seq(wall_description_schema),
     }
 )
-# [[[end]]] (checksum: 2b63e03d8d0f20fad052f0a1f63fa7d7)
+# [[[end]]] (checksum: ecc53a4621703890071145dadb78e781)
 # fmt: on