+++ tests/common_testing/alfasim_sdk_common_testing/filled_case_descriptions.py
@@ -52,7 +52,7 @@ PVT_MODEL_CORRELATION_DEFINITION = case_description.PvtModelCorrelationDescripti
     rs_sat=Scalar(4, "sm3/sm3"),
     pvt_correlation_package=constants.CorrelationPackage.Lasater,
 )
-FLUID_DESCRIPTION = case_description.FluidDescription(
+COMPOSITIONAL_FLUID_DESCRIPTION = case_description.CompositionalFluidDescription(
     composition=[COMPOSITION_DESCRIPTION_C1, COMPOSITION_DESCRIPTION_C2],
     fraction_pairs=[BIP_DESCRIPTION],
 )
@@ -65,16 +65,26 @@ PVT_MODEL_COMPOSITIONAL_DEFINITION = case_description.PvtModelCompositionalDescr
         LIGH_COMPONENT_DESCRIPTION,
         LIGH_COMPONENT_DESCRIPTION_OVERWRITE_C3,
     ],
-    fluids={"fluid_1": FLUID_DESCRIPTION},
+    fluids={"fluid_1": COMPOSITIONAL_FLUID_DESCRIPTION},
+)
+COMBINED_FLUID_DESCRIPTION = case_description.CombinedFluidDescription(
+    pvt_model="acme",
+)
+PVT_MODEL_COMBINED_DEFINITION = case_description.PvtModelCombinedDescription(
+    reference_pvt_model="acme",
+    fluids={"combined_fluid_1": COMBINED_FLUID_DESCRIPTION},
 )
 PVT_MODEL_TABLE_PARAMETERS = (
     case_description.PvtModelTableParametersDescription.create_constant(has_water=True)
 )
 PVT_MODELS_DEFINITION = case_description.PvtModelsDescription(
     default_model="acme",
-    compositions={
+    compositional={
         "composition 1": PVT_MODEL_COMPOSITIONAL_DEFINITION,
         "composition 2": PVT_MODEL_COMPOSITIONAL_DEFINITION,
+    },
+    combined={
+        "combined 1": PVT_MODEL_COMBINED_DEFINITION,
     },
     correlations={
         "correlation 1": PVT_MODEL_CORRELATION_DEFINITION,